refactor(routers): move catalog routes
