+++ docs/context/interface/api.md
@@ -4,6 +4,7 @@ status: shipped
 sources:
   - src/treg/api.py
   - src/treg/routers/__init__.py
+  - src/treg/routers/catalog.py
   - src/treg/routers/dependencies.py
   - src/treg/timeutil.py
   - src/treg/catalog_store.py
@@ -19,7 +20,9 @@ related:
 
 # The API
 
-Route definitions live on `api.router`; `bootstrap.create_app()` assembles them into FastAPI roles.
+Route definitions live on `api.router`; the open Catalog JSON block is defined in
+`routers.catalog` and attached at its original registration point. `bootstrap.create_app()` assembles
+the combined route table into FastAPI roles.
 `api.app` remains the deployed, backward-compatible `all` role. Everything the CLI + skill do is one
 HTTP call over this. The factory lifespan
 runs `init_db()`, then `_backfill_provider_extra_tools()` (the idempotent repair for provider registry
